Use cron to schedule items
This was WAY simpler than trying to parse a "day of week schedule" by hand. This simply parses the HH:mm text, converts it to a cron string, and schedules it. Another benefit is it doesn't need to continually schedule the next item, but everything gets scheduled recurringly as soon as the base server is started.