Refactor unit tests
Tests are now more isolated. Dependencies are all mocked and don't depend on their functionality. Also most of the thermostat tests were converted to mock the obserable directly off the temp reader, so we now have access to the completed callback of it. In this callback assertions can be made and done() can be called so there's no chance of emitted values bleeding into other test runs like before.