Properly convert string -> enum
The conversion was failing in the server, resulting in settings the mode to "0" or "1" instead of the actualy enum value. Fixed the test to properly test this (it was not testing right) and added a bit to the thermostat to send out an error on the IoT bridge if it gets anything other than the expected enum values.