Add one schedule per mode
Previously one schedule was used for both heating and cooling. In practice, this isn't ideal as the schedules for each of these modes will almost always be different. This adds the ability to have a separate schedule for both heating and cooling. The scheduler now takes a schedule object that contains both heating/cooling schedules and the server passes in the correct mode to pick. The server watches for mode changes and re-initializes the schedule when this happens.